cookify-ml/requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
idna==3.7
Jinja2==3.1.3
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pika==1.3.2
preshed==3.0.9
psycopg2==2.9.9
pt-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.7.0/pt_core_news_sm-3.7.0-py3-none-any.whl#sha256=a3a8eed42c600af17c70ed3c8720ffe2879b5353055e66fe0c998306329052fd
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
